14_Authentications/token_based_authentication/backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
PyJWT
python-dotenv
argon2-cffi
